+++ spinta/utils/enums.py
@@ -1,4 +1,10 @@
+from enum import Enum
+from typing import Any
+from typing import Optional
+from typing import Type
+
 from spinta import exceptions
+from spinta.components import Component
 
 
 def get_enum_by_name(enum, value):
@@ -8,7 +14,12 @@ def get_enum_by_name(enum, value):
     raise Exception(f"Unknown value {value!r}.")
 
 
-def enum_by_name(component, param, enum, name):
+def enum_by_name(
+    component: Component,
+    param: Optional[str],  # component param
+    enum: Type[Enum],
+    name: Any,
+) -> Optional[Enum]:
     if name is None or name == '':
         return None
     for item in enum: